Removed extra code files
