--- src/test/resources/mongopush/includeOptionsTests
@@ -1,4 +1,0 @@
-[{"namespace":"POCDB.POCCOLL","filter":{"fld0": {"$gt":500000}}}]
-[{"namespace":"POCDB.POCCOLL","filter":{"fld1":{"$date":"2016-11-04T05:57:25.910Z"}}}]
-[{"namespace":"POCDB.POCCOLL","filter":{"fld2":".*et justo"}}]
-[{"namespace":"POCDB.POCCOLL","filter":{"fld3":".* dolor"}}]
