--- src/test/resources/mongopush/pushDataTests
@@ -1,5 +1,0 @@
-POC_DRIVER_MONGO_PUSH_DATA_SYNC
-POC_DRIVER_MONGO_PUSH_DATA_ONLY_SYNC
-POC_DRIVER_MONGO_PUSH_DATA_ONLY_SYNC_VERIFY
-POC_DRIVER_MONGO_PUSH_DATA_SYNC_OPLOG
-POC_DRIVER_MONGO_PUSH_DATA_ONLY_SYNC_VERIFY_FAILED
